requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
django-appdata==0.1.3
django-filer
django-mptt==0.6.0
django-polymorphic==0.5.5
easy-thumbnails==2.0.1
gnureadline==6.3.3
ipdb==0.8
ipython==2.1.0
wsgiref==0.1.2
